+++ brightpath/__init__.py
@@ -1,5 +1,5 @@
 __all__ = ("BrightwayConverter", "SimaproConverter", "DATA_DIR")
-__version__ = (0, 0, 3)
+__version__ = (0, 0, 4)
 
 from pathlib import Path